backend/pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = ">=0.109.1,<1"
requests = "^2.32.0"
types-requests = "^2.31.0"
pydantic = "^2.10.2"
pyhumps = "^3.8.0"
uvicorn = ">=0.23.1,<1"
python-ulid = "^1.1.0"
python-jose = "^3.3.0"
boto3 = "^1.35.9"
pg8000 = "^1.30.3"
argparse = "^1.4.0"
tenacity = "<=8.3.0"
langdetect = "^1.0.9"
retry = ">=0.9.2,<1"
types-retry = ">=0.9.9.4,<1"
duckduckgo-search = "^7.3.0"
boto3-stubs = {extras = ["bedrock", "bedrock-agent-runtime", "bedrock-runtime", "boto3"], version = "^1.35.41"}

[tool.poetry.group.dev.dependencies]
mypy = "^1.10.0"
black = "^24.8.0"
appmap = "^2.1.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
